+++ src/mail/interfaces/mail.service.interface.ts
@@ -1,7 +1,5 @@
 export interface IMailService {
-  sendVerificationEmail(
-    to: string,
-    displayName: string,
-    rawToken: string,
-  ): Promise<void>;
+  sendVerificationEmail(to: string, displayName: string, rawToken: string): Promise<void>;
+  sendWelcomeEmail(to: string, displayName: string): Promise<void>;
+  sendCustomEmail(to: string, subject: string, content: string): Promise<void>;
 }